// src/app/enum/categorie-vehicule-enum.enum.ts
export enum CategorieVehiculeEnum {

  MICRO_URBAINES = "Micro-urbaines",
  MINI_CITADINES = "Mini-citadines",
  CITADINES_POLYVALENTES = "Citadines polyvalentes",
  COMPACTES = "Compactes",
  BERLINES_TAILLE_S = "Berlines Taille S",
  BERLINES_TAILLE_M = "Berlines Taille M",
  BERLINES_TAILLE_L = "Berlines Taille L",
  SUV_TOUTTERRAINS_PICKUP = "SUV, Tout-terrains, Pick-up"

}